requirements.txt
# Core Application & UI
streamlit>=1.20.0
streamlit-agraph>=0.0.47 # For graph visualization in Streamlit

# Code Parsing & AST
astor>=0.8.1 # For converting AST nodes back to source (consider ast.unparse for Py 3.9+)

# Graph Analysis
networkx>=3.0

# RAG & Vector Store
faiss-cpu>=1.7.4 # For vector indexing and search (CPU version)
numpy>=1.20.0

# LLM Interaction
ollama>=0.1.8 # Official Ollama Python client
requests>=2.25.0 # For making API calls (e.g., to DeepSeek)

# Configuration & Utilities
python-dotenv>=1.0.0 # For loading .env files

# Note: Ensure you have a C++ compiler available if faiss-cpu needs to build from source.
#       Pre-built wheels are available for many platforms.
